fix: address CodeQL review comments and add HikariCP interop verification
Review comments (github-code-quality):
- LocalFlightServerTest: widen to long before multiplication in row-count
  assertion
- TestFlightSqlServer: parse endpoint indexes defensively, converting
  malformed ticket handles into INVALID_ARGUMENT instead of an uncaught
  NumberFormatException (both getStream methods, shared helper)

HikariCP interop:
- New HikariCpCompatTest: HikariCP-pooled connections to Spice through
  the Arrow Flight SQL JDBC driver — plain and prepared statements,
  connection reuse, 4-thread concurrency, and authenticated pools using
  the same handshake credentials as the native client
- TestFlightSqlServer now advertises dataset + parameter schemas in
  CreatePreparedStatement results (JDBC/Avatica clients treat an empty
  dataset schema as an UPDATE and route execution to DoPut)
- README: connection pooling guidance (native SpiceClient needs no pool;
  HikariCP via Flight SQL JDBC for JDBC ecosystems); release notes updated
- Test-scope deps: HikariCP 5.1.0, flight-sql-jdbc-core 19.0.0

262 tests, 0 failures; SpotBugs clean.

+++ src/test/java/ai/spice/LocalFlightServerTest.java
@@ -124,7 +124,7 @@ public class LocalFlightServerTest extends TestCase {
     public void testPlainQueryConsumesOnlyFirstEndpoint() throws Exception {
         server.endpointCount = 3;
         try (FlightStream stream = client.query("SELECT * FROM test")) {
-            assertEquals(server.batchesPerEndpoint * server.rowsPerBatch, countRows(stream));
+            assertEquals((long) server.batchesPerEndpoint * server.rowsPerBatch, countRows(stream));
         }
         assertEquals(1, server.doGetCalls.get());
     }
